requirements.txt
flask==2.2.5
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu==1.12.0
numpy>=1.23.0
PyPDF2>=3.0.0
tqdm>=4.64.0
